requirements.txt
openai==1.12.0
httpx==0.26.0
python-dotenv==1.0.0
duckduckgo-search>=4.1.1 